move service_setting to make accessible in xpack
