+++ src/ui/public/vis/__tests__/map/service_settings.js
