+++ src/ui/public/vis/map/service_settings.js
